add selectedRow

+++ src/app/features/appointments/appointments.component.ts
@@ -1,4 +1,4 @@
-import { Component, ViewChild, inject } from "@angular/core";
+import { Component, Signal, ViewChild, WritableSignal, inject, signal } from "@angular/core";
 import { TranslateModule } from "@ngx-translate/core";
 import { MatTableDataSource, MatTableModule } from "@angular/material/table";
 import { MatIcon } from "@angular/material/icon";
@@ -7,11 +7,13 @@ import { MatCard } from "@angular/material/card";
 import { AppointmentsService } from "src/app/core/services/appointments.service";
 import { MatSort, MatSortModule } from "@angular/material/sort";
 import { formatDate, untilDestroyed } from "src/app/core/helpers/utils";
-import { AsyncPipe, DatePipe, NgFor } from "@angular/common";
+import { AsyncPipe, DatePipe, NgClass, NgFor } from "@angular/common";
 import { ProcessWheelComponent } from "src/app/shared/components/process-wheel/process-wheel.component";
 import { BehaviorSubject, Observable, Subject, async, finalize, map, of, take } from "rxjs";
 import { NotificationService } from "src/app/core/services/notification.service";
 import { ResolveStart } from "@angular/router";
+import { MatDialog } from "@angular/material/dialog";
+import { ConfirmationDialogComponent } from "src/app/shared/components/confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: "app-appointments",
@@ -26,6 +28,7 @@ import { ResolveStart } from "@angular/router";
     DatePipe,
     ProcessWheelComponent,
     NgFor,
+    NgClass
   ],
   templateUrl: "./appointments.component.html",
   styleUrl: "./appointments.component.scss",
@@ -40,9 +43,11 @@ export class AppointmentsComponent {
   allAppoitmentsAvailable: Subject<boolean> = new Subject();
   allallAppoitmentsAvailable$ = this.allAppoitmentsAvailable.asObservable();
   loading = false;
+  selectedRow: WritableSignal<Appointment | null> = signal(null)
 
   #apoitmentService = inject(AppointmentsService);
   #notificationService = inject(NotificationService);
+  #matDialogService = inject(MatDialog);
 
   ngOnInit() {
     this.getAllAppoitments();
@@ -78,9 +83,18 @@ export class AppointmentsComponent {
   }
 
   deleteAppointment(appointment: Appointment) {
-    this.#apoitmentService.deleteAppointment(appointment).subscribe((response) => {
-      this.#notificationService.showSuccess("Успешно изтрихте часа!");
-      this.getAllAppoitments();
-    });
+    this.selectedRow.set(appointment)
+    this.#matDialogService
+      .open(ConfirmationDialogComponent)
+      .afterClosed()
+      .subscribe((res) => {
+        this.selectedRow.set(null)
+        if (res) {
+          this.#apoitmentService.deleteAppointment(appointment).subscribe((response) => {
+            this.#notificationService.showSuccess("Успешно изтрихте часа!");
+            this.getAllAppoitments();
+          });
+        }
+      });
   }
 }
